backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-generativeai>=0.8.0
ddgs>=9.12.0
googlesearch-python>=1.3.0
pdfplumber>=0.11.0
python-docx>=1.1.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
python-multipart>=0.0.18
aiofiles>=24.1.0
